Initial support for custom kernel and tdnf
Outstanding:
* allow each config to supply it's own kernel parameters in isolinux.cfg

Introduces an isos/base/repos directory that holds the
configuration/specification details for a particular variant, including
package mangaer configuration, linux kernel version, and distro packages
required.

This change allows for selection of a specific kernel for the bootstrap
iso and adds initial Photon 2.0 support.

Switches to use udevd or systemd-udevd based based on the init subsystem
type.

Adds a basic breakout of the live init process after bootstrap boot and
makes the use of entropy daemon and iptables a little more generalized.

yum does not support rich dependencies so we will need to move to tdnf as
a package manager during the build in order to adopt photon 2.0 reliably.

Adds per repo configuration of iptables and entropy daemon to allow
customization based on package availability and kernel support.

Switches to building the ISOs in containers, based off either Photon2.0 or
Centos. This is done to get "native" tdnf or yum package managers -
necessary to work around quirks of the two tools.

This also adds convenience targets for the specific vic-machine variants
given a lot of our users use windows and it makes providing build
instructions simpler.

+++ isos/bootstrap.sh
@@ -34,21 +34,16 @@ tempfs_target_list=('/lib/modules/*' \
 # Build the bootstrap filesystem ontop of the base
 
 # exit on failure
-set -e
-
-if [ -n "$DEBUG" ]; then
-    set -x
-fi
-
+set -e && [ -n "$DEBUG" ] && set -x
 DIR=$(dirname $(readlink -f "$0"))
 . $DIR/base/utils.sh
 
 function usage() {
-echo "Usage: $0 -p staged-package(tgz) -b binary-dir -d <activates debug when set>" 1>&2
+echo "Usage: $0 -p staged-package(tgz) -b binary-dir" 1>&2
 exit 1
 }
 
-while getopts "p:b:d:" flag
+while getopts "p:b:o:" flag
 do
     case $flag in
 
@@ -61,9 +56,9 @@ do
             # Required. Target for iso and source for components
             BIN="$OPTARG"
             ;;
-        d)
-            # Optional. directs script to make a debug iso instead of a production iso.
-            debug="$OPTARG"
+        o)
+            # Optional. Name of the generated ISO
+            ISONAME="$OPTARG"
             ;;
         *)
 
@@ -88,34 +83,60 @@ PKGDIR=$(mktemp -d)
 
 unpack $package $PKGDIR
 
+# load the repo to use from the package if not explicit in env
+REPO=${REPO:-$(cat $PKGDIR/repo.cfg)}
+REPODIR="$DIR/base/repos/${REPO}/"
+PACKAGE_MANAGER=${PACKAGE_MANAGER:-$(cat $REPODIR/repo-spec.json | jq -r '.packagemanager')}
+PACKAGE_MANAGER=${PACKAGE_MANAGER:-tdnf}
+setup_pm $REPODIR $PKGDIR $PACKAGE_MANAGER $REPO
+
+
 #selecting the init script as our entry point.
-if [ -v debug ]; then
-    export ISONAME="bootstrap-debug.iso"
-    cp ${DIR}/bootstrap/bootstrap.debug $(rootfs_dir $PKGDIR)/bin/bootstrap
-    cp ${BIN}/rpctool $(rootfs_dir $PKGDIR)/sbin/
-else
-    export ISONAME="bootstrap.iso"
-    cp ${DIR}/bootstrap/bootstrap $(rootfs_dir $PKGDIR)/bin/bootstrap
-fi
+export ISONAME=${ISONAME:-bootstrap.iso}
+cp ${DIR}/bootstrap/bootstrap $(rootfs_dir $PKGDIR)/bin/bootstrap
 
 # copy in our components
 cp ${BIN}/tether-linux $(rootfs_dir $PKGDIR)/bin/tether
 cp ${BIN}/unpack $(rootfs_dir $PKGDIR)/bin/unpack
 
-# kick off our components at boot time
-mkdir -p $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants
-cp ${DIR}/bootstrap/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/
-cp ${DIR}/appliance/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/
-ln -s /etc/systemd/system/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants/
-ln -sf /etc/systemd/system/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/default.target
+if [ -d $(rootfs_dir $PKGDIR)/etc/systemd ]; then
+    echo "Preparing systemd for bootstrap"
 
-# disable networkd given we manage the link state directly
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/multi-user.target.wants/systemd-networkd.service
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/sockets.target.wants/systemd-networkd.socket
+    # copy in systemd entry script
+    cp ${DIR}/bootstrap/systemd-init $(rootfs_dir $PKGDIR)/bin/init
 
-# do not use the systemd dhcp client
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/network/*
-cp ${DIR}/base/no-dhcp.network $(rootfs_dir $PKGDIR)/etc/systemd/network/
+    # kick off our components at boot time
+    mkdir -p $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants
+    cp ${DIR}/bootstrap/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/
+    cp ${DIR}/appliance/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/
+    ln -s /etc/systemd/system/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants/
+    ln -sf /etc/systemd/system/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/default.target
+
+    # disable networkd given we manage the link state directly
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/multi-user.target.wants/systemd-networkd.service
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/sockets.target.wants/systemd-networkd.socket
+
+    # do not use the systemd dhcp client
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/network/*
+
+    # some systemd distros (centos-7) do not use systemd-networkd
+    [ -e $(rootfs_dir $PKGDIR)/etc/systemd/network/ ] && cp ${DIR}/base/no-dhcp.network $(rootfs_dir $PKGDIR)/etc/systemd/network/
+else
+    echo "Preparing systemV init for bootstrap"
+    
+    # copy in sysv-init entry script
+    cp ${DIR}/bootstrap/sysv-init $(rootfs_dir $PKGDIR)/bin/init
+
+    # kick off our components at boot time
+    cp ${DIR}/bootstrap/tether $(rootfs_dir $PKGDIR)/etc/rc.d/init.d/
+    chmod +x $(rootfs_dir $PKGDIR)/etc/rc.d/init.d/tether
+    ln -sf /etc/rc.d/init.d/tether $(rootfs_dir $PKGDIR)/etc/rc.d/rc1.d/S90tether
+
+    # set the default run level for sysvinit
+    # networking disabled on rc1
+    cp ${DIR}/bootstrap/inittab $(rootfs_dir $PKGDIR)/etc/inittab
+fi
+cp ${REPODIR}/init.sh $(rootfs_dir $PKGDIR)/bin/repoinit
 
 # compute the size of the target tempfs,
 # the list of directories/files in ${tempfs_target_list} should
@@ -131,4 +152,5 @@ echo Total tempfs size: ${size}
 echo "${tempfs_target_list[@]}" > $(rootfs_dir $PKGDIR)/.tempfs_list
 echo ${size} > $(rootfs_dir $PKGDIR)/.tempfs_size
 
-generate_iso $PKGDIR $BIN/$ISONAME /lib/systemd/systemd
+INIT=$(cat $REPODIR/repo-spec.json | jq -r '.init')
+generate_iso $PKGDIR $BIN/$ISONAME $INIT
